fixed rename add-offer component
