+++ src/components/account/add-offer/add-offer.tsx
@@ -7,7 +7,7 @@ import { Button } from "@src/shared/ui/button";
 import { Card } from "@src/components/account/card";
 import { ListItem } from "@src/components/account/card/list-item";
 
-import styles from "./addOffer.module.scss";
+import styles from "./add-offer.module.scss";
 import { customerOrders, CustomerOrdersItems } from "@src/shared/data/performer-account";
 import { Slider } from "@src/components/account/slider";
 import { useForm } from "react-hook-form";
