--- src/components/account/addOffer/index.ts
@@ -1,1 +1,0 @@
-export * from "./addOffer";
